251217_11_홍희표

+++ app/src/main/java/com/survivalcoding/gangnam2kiandroidstudy/core/routing/Route.kt
@@ -33,5 +33,5 @@ sealed interface Route : NavKey {
     data object Profile : Route
 
     @Serializable
-    data class RecipeDetails(val recipeId: Long) : Route
+    data class RecipeDetails(val recipeId: Int) : Route
 }
